machinelearningteam/summary-statistics-service/summary-service/requirements.txt
pandas==1.2.1
pytest
protobuf
grpcio-tools
mypy-protobuf
mypy